requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-filter==24.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
